-r,-a,-l のオプションに対応やり直し
